refactor(igx-grid): Reworked barrel imports
