+++ projects/igniteui-angular/src/lib/grids/common/column.interface.ts
