+++ projects/igniteui-angular/src/lib/grids/common/enums.ts
