+++ projects/igniteui-angular/src/lib/grids/common/templates.ts
