+++ projects/igniteui-angular/src/lib/grids/grid-common.module.ts
@@ -67,7 +67,7 @@ import {
     IgxCellEditorTemplateDirective,
     IgxCellTemplateDirective,
     IgxFilterCellTemplateDirective
-} from './types';
+} from './common/templates';
 import { IgxRowSelectorsModule } from './igx-row-selectors.module';
 /**
  * @hidden
