+++ projects/igniteui-angular/src/lib/grids/grid/grid-summary.spec.ts
@@ -26,7 +26,7 @@ import { setupGridScrollDetection, resizeObserverIgnoreError } from '../../test-
 import { SampleTestData } from '../../test-utils/sample-test-data.spec';
 import { IgxStringFilteringOperand, IgxNumberFilteringOperand, SortingDirection, IgxChipComponent } from 'igniteui-angular';
 import { ColumnGroupFourLevelTestComponent } from './column-group.spec';
-import { GridSummaryCalculationMode } from '../types';
+import { GridSummaryCalculationMode } from '../common/enums';
 
 describe('IgxGrid - Summaries #grid', () => {
     configureTestSuite();
