+++ projects/igniteui-angular/src/lib/grids/grid/grid.component.ts
@@ -11,12 +11,9 @@ import { cloneArray, IBaseEventArgs } from '../../core/utils';
 import { IGroupByRecord } from '../../data-operations/groupby-record.interface';
 import { IgxGroupByRowTemplateDirective } from './grid.directives';
 import { IgxGridGroupByRowComponent } from './groupby-row.component';
-// import { IDisplayDensityOptions, DisplayDensityToken } from '../../core/displayDensity';
 import { IGroupByExpandState } from '../../data-operations/groupby-expand-state.interface';
 import { IBaseChipEventArgs, IChipClickEventArgs, IChipKeyDownEventArgs } from '../../chips/chip.component';
 import { IChipsAreaReorderEventArgs } from '../../chips/chips-area.component';
-// import { TransactionService, Transaction, State } from '../../services/transaction/transaction';
-// import { DOCUMENT } from '@angular/common';
 import { IgxColumnComponent } from '../column.component';
 import { takeUntil } from 'rxjs/operators';
 import { IgxFilteringService } from '../filtering/grid-filtering.service';
@@ -24,11 +21,10 @@ import { IGroupingExpression } from '../../data-operations/grouping-expression.i
 import { IgxColumnResizingService } from '../grid-column-resizing.service';
 import { IgxGridSummaryService } from '../summaries/grid-summary.service';
 import { IgxGridSelectionService, IgxGridCRUDService } from '../../core/grid-selection';
-// import { IgxOverlayService } from '../../services/index';
 import { IgxForOfSyncService } from '../../directives/for-of/for_of.sync.service';
 import { IgxDragIndicatorIconDirective } from '../row-drag.directive';
 import { IgxGridMRLNavigationService } from '../grid-mrl-navigation.service';
-import { FilterMode } from '../types';
+import { FilterMode } from '../common/enums';
 
 let NEXT_ID = 0;
 
@@ -220,30 +216,6 @@ export class IgxGridComponent extends IgxGridBaseComponent implements IGridDataB
         return this.gridAPI as IgxGridAPIService;
     }
     private _filteredData = null;
-
-    // constructor(
-    //     selectionService: IgxGridSelectionService,
-    //     crudService: IgxGridCRUDService,
-    //     public colResizingService: IgxColumnResizingService,
-    //     gridAPI: GridBaseAPIService<IgxGridBaseComponent & IGridDataBindable>,
-    //     @Inject(IgxGridTransaction) _transactions: TransactionService<Transaction, State>,
-    //     elementRef: ElementRef,
-    //     zone: NgZone,
-    //     @Inject(DOCUMENT) public document,
-    //     cdr: ChangeDetectorRef,
-    //     resolver: ComponentFactoryResolver,
-    //     differs: IterableDiffers,
-    //     viewRef: ViewContainerRef,
-    //     navigation: IgxGridNavigationService,
-    //     filteringService: IgxFilteringService,
-    //     @Inject(IgxOverlayService) protected overlayService: IgxOverlayService,
-    //     summaryService: IgxGridSummaryService,
-    //     @Optional() @Inject(DisplayDensityToken) protected _displayDensityOptions: IDisplayDensityOptions) {
-    //         super(selectionService,
-    //               crudService, gridAPI, _transactions, elementRef, zone, document, cdr, resolver, differs, viewRef, navigation,
-    //               filteringService, overlayService, summaryService, _displayDensityOptions);
-    //         this._gridAPI = <IgxGridAPIService>gridAPI;
-    // }
 
     /**
      * Returns the group by state of the `IgxGridComponent`.
