+++ projects/igniteui-angular/src/lib/grids/grid/grid.groupby.spec.ts
@@ -18,7 +18,7 @@ import { DataParent } from '../../test-utils/sample-test-data.spec';
 import { MultiColumnHeadersWithGroupingComponent } from '../../test-utils/grid-samples.spec';
 import { resizeObserverIgnoreError, HelperUtils } from '../../test-utils/helper-utils.spec';
 import { GridSelectionFunctions } from '../../test-utils/grid-functions.spec';
-import { GridSelectionMode } from '../types';
+import { GridSelectionMode } from '../common/enums';
 
 describe('IgxGrid - GroupBy #grid', () => {
     configureTestSuite();
