+++ projects/igniteui-angular/src/lib/grids/tree-grid/tree-grid-expanding.spec.ts
@@ -17,7 +17,7 @@ import { wait } from '../../test-utils/ui-interactions.spec';
 import { IgxGridModule } from '../grid';
 import { resizeObserverIgnoreError } from '../../test-utils/helper-utils.spec';
 import { GridFunctions } from '../../test-utils/grid-functions.spec';
-import { GridSelectionMode } from '../types';
+import { GridSelectionMode } from '../common/enums';
 
 describe('IgxTreeGrid - Expanding / Collapsing #tGrid', () => {
     configureTestSuite();
