--- projects/igniteui-angular/src/lib/grids/types/index.ts
@@ -1,5 +1,0 @@
-export * from './templates';
-export * from './enums';
-export * from './grid.interface';
-export * from './column.interface';
-export * from './row.interface';
